Remove ILA; Pcam Fixes
Remove debugging ILA and fix AXI FIFOs for Pcam. In order to
account for improper AXI flushing on reset by the framebuffer
write IP the AXI FIFO must be reset along with it. Importantly,
the AXI interconnect must also not have any FIFOs enabled. Failure
to do this causes the image to become misaligned after the first
time the framebuffer write IP is disabled. It seems this has been
fixed in future versions of the IP, so after update the AXI
FIFO can likely be removed (and FIFOs reenabled at the interconnect)